Only allow projects younger than site deprecation date
Short-term fix that only applies /after/ fetching the data. Ideal solution is to apply filter in the actual query, before any requests occur (why ask for a project if you're not even going to end up using it--cut down on unnecessary data transfers!) // The architecture (that I made as a result of keeping an editorial mindset when creating the re-ordering options) makes it difficult to do easily though. Definitely on my list to address or research to improve next time around.